+++ CHANGES.rst
@@ -11,6 +11,12 @@
   memory corruption could result. See `issue 245
   <https://github.com/python-greenlet/greenlet/issues/245>`_. Patch by
   fygao-wish.
+- Fix a leak of a list object when the last reference to a greenlet
+  was deleted from some other thread than the one it belonged too. For
+  this to work correctly, you must call a greenlet API like
+  ``getcurrent()`` before the thread owning the greenlet exits; we
+  hope to lift this limitation. Note that in some cases this may also
+  fix leaks of greenlet objects themselves. See `issue 251 <>`_.
 
 
 1.1.1 (2021-08-06)